feature: doris instance detail state
